refactor: 画像を表示する処理をShowImageとして切り分けた
