+++ GourmetApp/View/ViewController/RestaurantDetailViewController.swift
@@ -57,30 +57,7 @@ final class RestaurantDetailViewController: UIViewController {
             subCatchLabel.text = data.shopCatch
             genreLabel.text = data.genre.name
             openLabel.text = data.open
-            showImage(imageUrl: data.photo.mobile.l)
+            ShowImage.showImage(imageUrl: data.photo.mobile.l, imageView: shopImage)
         }
     }
-
-    private func showImage(imageUrl: String) {
-        guard let url = URL(string: imageUrl) else {
-            return
-        }
-
-        URLSession.shared.dataTask(with: url) { [weak self] data, _, error in
-            if let error = error {
-                print(error)
-                return
-            }
-
-            if let data = data,
-               let image = UIImage(data: data){
-                DispatchQueue.main.async {
-                    self?.shopImage.image = image
-                }
-            } else {
-                print("画像表示エラー")
-            }
-        }.resume()
-    }
-
 }
